refactor: Изменение файловой структуры тестов
