+++ src/components/__tests__/App.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import App from './App';
+import App from '../App';
 
 test('Приложение рендерится без падения', () => {
   const app = shallow(<App />)
